requirements.txt
asgiref==3.3.1
Django==3.0
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2021.1
sqlparse==0.4.1
